interformer:reupdate missing files

+++ interformer/model/transformer/graphormer/interformer.py
@@ -191,9 +191,9 @@ class VinaScoreHead(nn.Module):
                 print(f"[Interformer] Gussian Score ->{target}-{last_id}")
                 data = {'ligand_len': batched_data['ligand_len'].cpu().numpy(),
                         'pocket_len': batched_data['pocket_len'].cpu().numpy(),
-                        'pi': pi_soft.cpu().float().numpy().squeeze(0),
-                        'mean': mean.cpu().float().numpy().squeeze(0),
-                        'sigma': sigma.cpu().float().numpy().squeeze(0),
+                        'pi': pi_soft.cpu().numpy().squeeze(0),
+                        'mean': mean.cpu().numpy().squeeze(0),
+                        'sigma': sigma.cpu().numpy().squeeze(0),
                         'hbond_pair': hbond_pair.cpu().numpy().squeeze(0),
                         'hydro_pair': hydro_pair.cpu().numpy().squeeze(0),
                         'd': d.cpu().numpy().squeeze(0),
@@ -349,6 +349,7 @@ class Interformer(SBDD):
         if self.pose_sel_mode:
             pose_logits = self.out_pose_sel_proj(vn_node)
             final.append(pose_logits)
+        # MARK: Final=[affinity, pose_logits]
         elif self.energy_mode:
             # Step6, G-Score
             _, gscore_loss, gscore_pos_loss = self.VinaScoreHead(output_node, output_edge, batched_data)
@@ -358,7 +359,7 @@ class Interformer(SBDD):
             # Predict Atom Type
             atom_loss = self.AtomTypeHead(output_node, batched_data['x'])
             final.append(atom_loss)
-        # MARK: Final=[affinity, pose_selection, gscore_loss, gscore_pos_loss, atom_loss]
+        # MARK: Final=[affinity, gscore_loss, gscore_pos_loss, atom_loss]
         return final
 
     @staticmethod
